修改expot文件为int

+++ export_parent/export_domain/src/main/java/cn/itcast/domain/cargo/ExportExample.java
@@ -156,7 +156,7 @@ public class ExportExample implements Serializable {
      *
      * @mbg.generated
      */
-    protected abstract static class GeneratedCriteria implements Serializable{
+    protected abstract static class GeneratedCriteria implements Serializable {
         protected List<Criterion> criteria;
 
         protected GeneratedCriteria() {
@@ -1217,7 +1217,7 @@ public class ExportExample implements Serializable {
             return (Criteria) this;
         }
 
-        public Criteria andStateEqualTo(Long value) {
+        public Criteria andStateEqualTo(Integer value) {
             addCriterion("state =", value, "state");
             return (Criteria) this;
         }
@@ -1747,7 +1747,7 @@ public class ExportExample implements Serializable {
      *
      * @mbg.generated
      */
-    public static class Criterion implements Serializable{
+    public static class Criterion implements Serializable {
         private String condition;
 
         private Object value;
